PING-Docker/python/requirements.txt
pandas
requests
numpy
mysql-connector-python
fuzzywuzzy
xlrd
